feat: move system to own file
