+++ src/lib.rs
@@ -17,7 +17,7 @@
 pub use crate::compression::{CompressionMethod, SUPPORTED_COMPRESSION_METHODS};
 pub use crate::datetime::DateTime;
 pub use crate::format::aes::AesMode;
-pub use crate::format::flags::System;
+pub use crate::format::system::System;
 pub use crate::format::{ZIP64_BYTES_THR, ZIP64_ENTRY_THR};
 pub use crate::read::HasZipMetadata;
 pub use crate::read::{ZipArchive, ZipReadOptions};
