+++ src/read/mod.rs
@@ -5,8 +5,8 @@ use crate::cp437::FromCp437;
 use crate::datetime::DateTime;
 use crate::extra_fields::{ExtraField, ExtraFields};
 use crate::format::blocks::{CentralDirectoryEndInfo, DataAndPosition, ZipCentralEntryBlock};
-use crate::format::flags::ZipFlags;
-use crate::format::flags::{System, ZipFileFlags};
+use crate::format::flags::{ZipFileFlags, ZipFlags};
+use crate::format::system::System;
 use crate::result::{ZipError, ZipResult, invalid};
 use crate::types::ZipFileData;
 use indexmap::IndexMap;
