+++ src/read/zipfile.rs
@@ -6,8 +6,8 @@ use crate::HasZipMetadata;
 use crate::ZIP64_BYTES_THR;
 use crate::ZipReadOptions;
 use crate::format::ffi;
-use crate::format::flags::System;
 use crate::format::flags::ZipFileFlags;
+use crate::format::system::System;
 use crate::read::ExtraField;
 use crate::read::RootDirFilter;
 use crate::read::make_writable_dir_all;
